+++ src/models/Recipe.php
@@ -12,10 +12,13 @@
 namespace nystudio107\recipe\models;
 
 use nystudio107\recipe\helpers\Json;
+use nystudio107\recipe\helpers\PluginTemplate;
 
 use Craft;
 use craft\base\Model;
 use craft\helpers\Template;
+
+use Twig\Markup;
 
 /**
  * @author    nystudio107
@@ -24,6 +27,36 @@ use craft\helpers\Template;
  */
 class Recipe extends Model
 {
+    // Constants
+    // =========================================================================
+
+    const US_RDA = [
+        'calories' => 2000,
+        'carbohydrateContent' => 275,
+        'cholesterolContent' => 300,
+        'fatContent' => 78,
+        'fiberContent' => 28,
+        'proteinContent' => 50,
+        'saturatedFatContent' => 20,
+        'sodiumContent' => 2300,
+        'sugarContent' => 50,
+    ];
+
+    // Mapping to convert any of the incorrect plural values
+    const NORMALIZE_PLURALS = [
+        'tsps' => 'tsp',
+        'tbsps' => 'tbsp',
+        'flozs' => 'floz',
+        'cups' => 'c',
+        'ozs' => 'oz',
+        'lbs' => 'lb',
+        'mls' => 'ml',
+        'ls' => 'l',
+        'mgs' => 'mg',
+        'gs' => 'g',
+        'kg' => 'kg',
+    ];
+
     // Public Properties
     // =========================================================================
 
@@ -148,6 +181,23 @@ class Recipe extends Model
     /**
      * @inheritdoc
      */
+    public function init()
+    {
+        parent::init();
+        // Fix any of the incorrect plural values
+        if (!empty($this->ingredients)) {
+            foreach ($this->ingredients as &$row) {
+                if (!empty($row['units']) && !empty(self::NORMALIZE_PLURALS[$row['units']])) {
+                    $row['units'] = self::NORMALIZE_PLURALS[$row['units']];
+                }
+            }
+            unset($row);
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function rules()
     {
         return [
@@ -185,19 +235,19 @@ class Recipe extends Model
     public function renderRecipeJSONLD($raw = true)
     {
         $recipeJSONLD = [
-            "context" => "http://schema.org",
-            "type" => "Recipe",
-            "name" => $this->name,
-            "image" => $this->getImageUrl(),
-            "description" => $this->description,
-            "recipeYield" => $this->serves,
-            "recipeIngredient" => $this->getIngredients("imperial", 0, false),
-            "recipeInstructions" => $this->getDirections(false),
+            'context' => 'http://schema.org',
+            'type' => 'Recipe',
+            'name' => $this->name,
+            'image' => $this->getImageUrl(),
+            'description' => $this->description,
+            'recipeYield' => $this->serves,
+            'recipeIngredient' => $this->getIngredients('imperial', 0, false),
+            'recipeInstructions' => $this->getDirections(false),
         ];
         $recipeJSONLD = array_filter($recipeJSONLD);
 
         $nutrition = [
-            "type" => "NutritionInformation",
+            'type' => 'NutritionInformation',
             'servingSize' => $this->servingSize,
             'calories' => $this->calories,
             'carbohydrateContent' => $this->carbohydrateContent,
@@ -213,13 +263,13 @@ class Recipe extends Model
         ];
         $nutrition = array_filter($nutrition);
         $recipeJSONLD['nutrition'] = $nutrition;
-        if (count($recipeJSONLD['nutrition']) == 1) {
+        if (count($recipeJSONLD['nutrition']) === 1) {
             unset($recipeJSONLD['nutrition']);
         }
         $aggregateRating = $this->getAggregateRating();
         if ($aggregateRating) {
             $aggregateRatings = [
-                "type" => "AggregateRating",
+                'type' => 'AggregateRating',
                 'ratingCount' => $this->getRatingsCount(),
                 'bestRating' => '5',
                 'worstRating' => '1',
@@ -231,31 +281,31 @@ class Recipe extends Model
             $reviews = [];
             foreach ($this->ratings as $rating) {
                 $review = [
-                    "type" => "Review",
+                    'type' => 'Review',
                     'author' => $rating['author'],
-                    'name' => $this->name." ".Craft::t("recipe", "Review"),
+                    'name' => $this->name . ' ' . Craft::t('recipe', 'Review'),
                     'description' => $rating['review'],
                     'reviewRating' => [
-                        "type" => "Rating",
+                        'type' => 'Rating',
                         'bestRating' => '5',
                         'worstRating' => '1',
                         'ratingValue' => $rating['rating'],
                     ],
                 ];
-                array_push($reviews, $review);
+                $reviews[] = $review;
             }
             $reviews = array_filter($reviews);
             $recipeJSONLD['review'] = $reviews;
         }
 
         if ($this->prepTime) {
-            $recipeJSONLD['prepTime'] = "PT".$this->prepTime."M";
+            $recipeJSONLD['prepTime'] = 'PT' . $this->prepTime . 'M';
         }
         if ($this->cookTime) {
-            $recipeJSONLD['cookTime'] = "PT".$this->cookTime."M";
+            $recipeJSONLD['cookTime'] = 'PT' . $this->cookTime . 'M';
         }
         if ($this->totalTime) {
-            $recipeJSONLD['totalTime'] = "PT".$this->totalTime."M";
+            $recipeJSONLD['totalTime'] = 'PT' . $this->totalTime . 'M';
         }
 
         return $this->renderJsonLd($recipeJSONLD, $raw);
@@ -264,19 +314,37 @@ class Recipe extends Model
     /**
      * Get the URL to the recipe's image
      *
+     * @param null $transform
+     *
      * @return null|string
      */
-    public function getImageUrl()
+    public function getImageUrl($transform = null)
     {
-        $result = "";
+        $result = '';
         if (isset($this->imageId) && $this->imageId) {
             $image = Craft::$app->getAssets()->getAssetById($this->imageId[0]);
             if ($image) {
-                $result = $image->url;
+                $result = $image->getUrl($transform);
             }
         }
 
         return $result;
+    }
+
+    /**
+     * Render the Nutrition Facts template
+     *
+     * @param array $rda
+     * @return Markup
+     */
+    public function renderNutritionFacts(array $rda = self::US_RDA): Markup {
+        return PluginTemplate::renderPluginTemplate(
+            'recipe-nutrition-facts',
+            [
+                'value' => $this,
+                'rda' => $rda,
+            ]
+        );
     }
 
     /**
@@ -288,14 +356,14 @@ class Recipe extends Model
      *
      * @return array
      */
-    public function getIngredients($outputUnits = "imperial", $serving = 0, $raw = true)
+    public function getIngredients($outputUnits = 'imperial', $serving = 0, $raw = true): array
     {
         $result = [];
-        
-        if ($this->ingredients != '') {
+
+        if (!empty($this->ingredients)) {
             foreach ($this->ingredients as $row) {
-                $convertedUnits = "";
-                $ingredient = "";
+                $convertedUnits = '';
+                $ingredient = '';
                 if ($row['quantity']) {
                     // Multiply the quantity by how many servings we want
                     $multiplier = 1;
@@ -305,62 +373,65 @@ class Recipe extends Model
                     $quantity = $row['quantity'] * $multiplier;
                     $originalQuantity = $quantity;
 
-                    // Do the units conversion
-
-                    if ($outputUnits == 'imperial') {
-                        if ($row['units'] == "mls") {
-                            $convertedUnits = "tsps";
-                            $quantity = $quantity * 0.2;
-                        }
-
-                        if ($row['units'] == "ls") {
-                            $convertedUnits = "cups";
-                            $quantity = $quantity * 4.2;
-                        }
-
-                        if ($row['units'] == "mgs") {
-                            $convertedUnits = "ozs";
-                            $quantity = $quantity * 0.000035274;
-                        }
-
-                        if ($row['units'] == "gs") {
-                            $convertedUnits = "ozs";
-                            $quantity = $quantity * 0.035274;
+                    // Do the imperial->metric units conversion
+                    if ($outputUnits === 'imperial') {
+                        switch ($row['units']) {
+                            case 'ml':
+                                $convertedUnits = 'tsp';
+                                $quantity *= 0.2;
+                                break;
+                            case 'l':
+                                $convertedUnits = 'c';
+                                $quantity *= 4.2;
+                                break;
+                            case 'mg':
+                                $convertedUnits = 'oz';
+                                $quantity *= 0.000035274;
+                                break;
+                            case 'g':
+                                $convertedUnits = 'oz';
+                                $quantity *= 0.035274;
+                                break;
+                            case 'kg':
+                                $convertedUnits = 'lb';
+                                $quantity *= 2.2046226218;
+                                break;
                         }
                     }
-
-                    if ($outputUnits == 'metric') {
-                        if ($row['units'] == "tsps") {
-                            $convertedUnits = "mls";
-                            $quantity = $quantity * 4.929;
-                        }
-
-                        if ($row['units'] == "tbsps") {
-                            $convertedUnits = "mls";
-                            $quantity = $quantity * 14.787;
-                        }
-
-                        if ($row['units'] == "flozs") {
-                            $convertedUnits = "mls";
-                            $quantity = $quantity * 29.574;
-                        }
-
-                        if ($row['units'] == "cups") {
-                            $convertedUnits = "ls";
-                            $quantity = $quantity * 0.236588;
-                        }
-
-                        if ($row['units'] == "ozs") {
-                            $convertedUnits = "gs";
-                            $quantity = $quantity * 28.3495;
+                    // Do the metric->imperial units conversion
+                    if ($outputUnits === 'metric') {
+                        switch ($row['units']) {
+                            case 'tsp':
+                                $convertedUnits = 'ml';
+                                $quantity *= 4.929;
+                                break;
+                            case 'tbsp':
+                                $convertedUnits = 'ml';
+                                $quantity *= 14.787;
+                                break;
+                            case 'floz':
+                                $convertedUnits = 'ml';
+                                $quantity *= 29.574;
+                                break;
+                            case 'c':
+                                $convertedUnits = 'l';
+                                $quantity *= 0.236588;
+                                break;
+                            case 'oz':
+                                $convertedUnits = 'g';
+                                $quantity *= 28.3495;
+                                break;
+                            case 'lb':
+                                $convertedUnits = 'kg';
+                                $quantity *= 0.45359237;
+                                break;
                         }
 
                         $quantity = round($quantity, 1);
                     }
 
                     // Convert imperial units to nice fractions
-
-                    if ($outputUnits == 'imperial') {
+                    if ($outputUnits === 'imperial') {
                         $quantity = $this->convertToFractions($quantity);
                     }
                     $ingredient .= $quantity;
@@ -374,16 +445,16 @@ class Recipe extends Model
                             $units = rtrim($units);
                             $units = rtrim($units, 's');
                         }
-                        $ingredient .= " ".$units;
+                        $ingredient .= ' ' . $units;
                     }
                 }
                 if ($row['ingredient']) {
-                    $ingredient .= " ".$row['ingredient'];
+                    $ingredient .= ' ' . $row['ingredient'];
                 }
                 if ($raw) {
                     $ingredient = Template::raw($ingredient);
                 }
-                array_push($result, $ingredient);
+                $result[] = $ingredient;
             }
         }
 
@@ -400,70 +471,60 @@ class Recipe extends Model
     private function convertToFractions($quantity)
     {
         $whole = floor($quantity);
-        $fraction = $quantity - $whole;
+        // Round the mantissa so we can do a floating point comparison without
+        // weirdness, per: https://www.php.net/manual/en/language.types.float.php#113703
+        $fraction = round($quantity - $whole, 3);
         switch ($fraction) {
             case 0:
-                $fraction = "";
+                $fraction = '';
                 break;
-
             case 0.25:
-                $fraction = " &frac14;";
+                $fraction = ' &frac14;';
                 break;
-
             case 0.33:
-                $fraction = " &frac13;";
+                $fraction = ' &frac13;';
                 break;
-                
             case 0.66:
-                $fraction = " &frac23;";
+                $fraction = ' &frac23;';
                 break;
-            
             case 0.165:
-                $fraction = " &frac16;";
+                $fraction = ' &frac16;';
                 break;
-
             case 0.5:
-                $fraction = " &frac12;";
+                $fraction = ' &frac12;';
                 break;
-
             case 0.75:
-                $fraction = " &frac34;";
+                $fraction = ' &frac34;';
                 break;
-
             case 0.125:
-                $fraction = " &#x215B;";
+                $fraction = ' &#x215B;';
                 break;
-
             case 0.375:
-                $fraction = " &#x215C;";
+                $fraction = ' &#x215C;';
                 break;
-
             case 0.625:
-                $fraction = " &#x215D;";
+                $fraction = ' &#x215D;';
                 break;
-
             case 0.875:
-                $fraction = " &#x215E;";
+                $fraction = ' &#x215E;';
                 break;
-
             default:
                 $precision = 5;
                 $pnum = round($fraction, $precision);
-                $denominator = pow(10, $precision);
+                $denominator = 10 ** $precision;
                 $numerator = $pnum * $denominator;
-                $fraction = "<sup>"
+                $fraction = '<sup>'
                     .$numerator
-                    ."</sup>&frasl;<sub>"
+                    . '</sup>&frasl;<sub>'
                     .$denominator
-                    ."</sub>";
+                    . '</sub>';
                 break;
         }
-        if ($whole == 0) {
-            $whole = "";
+        if ($whole === 0) {
+            $whole = '';
         }
-        $result = $whole.$fraction;
 
-        return $result;
+        return $whole.$fraction;
     }
 
     /**
@@ -476,13 +537,13 @@ class Recipe extends Model
     public function getDirections($raw = true)
     {
         $result = [];
-        if ($this->directions != '') {
+        if (!empty($this->directions)) {
             foreach ($this->directions as $row) {
                 $direction = $row['direction'];
                 if ($raw) {
                     $direction = Template::raw($direction);
                 }
-                array_push($result, $direction);
+                $result[] = $direction;
             }
         }
 
@@ -503,26 +564,26 @@ class Recipe extends Model
                 $result += $row['rating'];
                 $total++;
             }
-            $result = $result / $total;
+            $result /= $total;
         } else {
-            $result = "";
+            $result = '';
         }
 
         return $result;
     }
-
-    // Private Methods
-    // =========================================================================
 
     /**
      * Get the total number of ratings
      *
      * @return int
      */
-    public function getRatingsCount()
+    public function getRatingsCount(): int
     {
         return count($this->ratings);
     }
+
+    // Private Methods
+    // =========================================================================
 
     /**
      * Renders a JSON-LD representation of the schema
@@ -534,7 +595,7 @@ class Recipe extends Model
      */
     private function renderJsonLd($json, $raw = true)
     {
-        $linebreak = "";
+        $linebreak = '';
 
         // If `devMode` is enabled, make the JSON-LD human-readable
         if (Craft::$app->getConfig()->getGeneral()->devMode) {